Implement and test DHT.run_coroutine
